zetta.py: implemented create function
  Now the notes creating functionality is available!
You can run `zetta create`. Zetta will place a new note
in the repo and call $EDITOR against it. After you quit
the editor, zetta will ask you to commit changes. The commit
message is the same as the first line of note (which is likely
to be its title).